cells: Implement escaping and quoting in argument parser
The argument parser of cells had rudimentary support for qouting, but
not for escaping. This means it was not possible to include quote
symbols within quoted strings, among other things.

This patch replaces the argument parser of cells. The new parser is a
recursive decent parser. It implements single and double quoting,
backslash escaping and the end of options double dash sequence.

The toString method of Args is updated to escape special characters.
This makes it possible to parse the output of toString and obtain an
equivalent instance of Args.

This patch is a prerequisite to another patch for CellShell that will
make it easier to include spaces in cell arguments.

Testing done: Verified that dCache starts and that transfer tests
succeed. Unit tests pass.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1913/