dcap: Fix duplicate response generation for mkdir
fileAttributesNotAvailable is called when the initial file attributes
lookup in PNFS fails. The method returns a boolean. When true,
fileAttributesAvailable will be called, assuming that the method created
the missing entry. If false is returned, the PnfsSessionHandler would
generate an error reply.

For MkDirHandler this behaviour didn't work: It needs to create the
directory in fileAttributesNotAvailable and depending on whether that
succeeds it must generate either an success or an error response. It
did, however when returning the generic code in PnfsSessionHandler would
generate a second response.

This patch changes the logic such that if false is returned, then it is
assumed that fileAttributesNotAvailable has generated a response
already.

The second problem fixed by this patch is that
MkDirHandler.fileAttributesNotAvailbable would return true in case of
failure. However that means that MkDirHandler.fileAttributesAvailable
would be called, generating a second error message. This is fixed by
always returning false.

I checked stable branches, and the bug is present in 1.9.5 and possibly
earlier versions. I suggest we do not fix the bug in stable branches.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1884/