login: Add login service
The login cell replaces the gPlazma cell. A login cell uses a
LoginStrategy internally. It can be defined with different types of
LogiStrategies, but we currently only have one with
GplazmaLoginStrategy.

A LoginMessage is introduced to request login operations.
RemoteLoginStrategy uses a login cell for login by sending it a
LoginMessage.

In constrast to the gPlazma cell, the login cell uses regular dCache
Message subclasses. Thus the client stubs are much simpler and error
handling is done by CellStub. Since login cell precisely matches the
abstraction level of LoginStrategy, implementing RemoteLoginStrategy is
trivial.

GplazmaLoginStrategy is changed from using a remote gplazma cell to
using gPlazma directly as a library. In contrast to the gPlazma cell,
GplazmaLoginStrategy doesn't support storing the login in a database. As
far as I could see, the data in the database of gPlazma was never
actually used (in contrast to the data stored by the SRM).

Like for LoginStrategy, only the login operation has been implemented.
The structure allows the addition of map and reverse once we add those
to the LoginStrategy interface.

The gplazma service definition is updated to instantiate a login cell
with a GplazmaLoginStrategy. There are no configuration changes.

Timeout handling is changed to rely on the TTL field in CellMessage
rather than a preconfigured timeout in the cell.

Eventually doors should be updated to make the login service to use be 
configurable (i.e. it should be possible to give the name of the cell it 
should contact).

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1936/