ftp: Check file type before processing NLST and MLSD
The FTP commands NLST and MLSD only apply to directories. For files,
these commands must cause an error to be reported.

RFC 3659 seems to imply that we have to report these errors before
opening the data channel for the listing. Currently we open the data
channel first and fail the listing afterwards. Besides causing a delay
for the case where the path is a file rather than a directory or when
the path doesn't even exist, our current behaviour breaks compatibility
with ARC clients. ARC clients fall back to LIST or MLST in case NLST and
MLSD report that an illegal argument was specified. However if the error
occurs after the data channel is opened, then the client assumes
something else failed and do not fall back.

Since the problem is caused by us violating the spec, this patch adds an
additional check before the data channel is opened.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1934/