pool: Recover acceptor thread in case of failures
We have observed that the P2P acceptor thread sometimes dies. This may
in particular happen when the file descriptor limit was reached, but
could also happen in case of other problems accepting a new connection.

This is a minimal patch to let the P2P acceptor thread be restarted in
case it dies due to some failure. The following changes are made:

- In start() we now check whether the worker thread is actually alive.
  If not we start it.

- In register() we call start() unconditionally. start() already has logic
  to avoid starting the worker thread twice.

- In the worker thread, we close the server socket upon termination to avoid
  leaking it.

Eventually the acceptor should be cleaned up on trunk. Since this is
however a bug that has force us to restart a number of pools, I propose
this minimal patch for stable branches.

Target: trunk
Target: 1.9.8 (proposed)
Target: 1.9.7 (proposed)
Target: 1.9.6 (proposed)
Target: 1.9.5 (proposed)
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1871/