httpd: Restructures batch file to allow site specific content override
The batch file is restructured to execute etc/httpd.conf if it is
defined. This allows the site to redefine the cell contexts that define
the content of the httpd service.

For the common case of adding new aliases rather than replacing the
existing content, etc/httpd.conf may also define the httpdSetupExtra
context.

The new batch file enables PoolInfoObserverV2 by default (I don't see a
reason not to have it).

The index page is updated with a link to the Pool Group view.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1942/