dcap: Remove permission handler checks
The DCAP door was the only door left still having the option to perform
permission checks in the door. All other doors rely exclusively on
PnfsManager for permission checks on name space operations.

This patch remove the permission checks in the DCAP door and the
configuration settings belonging to it.

As no doors are left using the PermissionHandler and the implementation
of the interface, the patch also remove those, together with the
associated unit tests.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1886/