AuthorizationRecord: Fix LoginReply to AuthorizationRecord conversion
DCacheAuthorization converts LoginReply to AuthorizationRecord. It
relied on the Subjects class for converting the Subject to
AuthorizationRecord and then it added the LoginAttributes afterwards.
Unfortunately an AuthorizationRecord may not be modified once its id
field has been computed. Thus the old logic for converting LoginReply to
AuthorizationRecord is broken.

This patch adds two constructors to AuthorizationRecord. One for
converting from Subject and another for converting from LoginReply. This
resolves the issue described above.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1955/