srm: Fix bug that causes configured queue limits to be ignored
The Scheduler component of the SRM uses three ModifiableQueue instances.
Each queue has a capacity. Unfortunately the limits set via setters are
not propagated to the queues and thus ignored.

There is a bigger structural problem in the Scheduler in that the limits
are partially handled in the Scheduler class and partially in the
ModifiableQueue. Thus the info output of the srm showed the expected
limits, but not the actual limits. This should be cleaned up at some
point.

The present patch is a minimal patch to resolve the bug. It does not
resolve the fact that there are multiple representations of the same
limits.

Since the bug makes it impossible to raise the TURL ready queue limits, 
I proposed the patch is merged into stable branches and new releases are
made ASAP. NDGF was down for some hours tonight because a client 
misbehaved and the default limit (1000) wasn't high enough to cope 
with the problem.

Target: trunk
Target: 1.9.8 (proposed)
Target: 1.9.7 (proposed)
Target: 1.9.6 (proposed)
Target: 1.9.5 (proposed)
Acked-by: Dmitry Litvinsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/1945/