PnfsManager: Implement default values for createEntry call
This patch makes it possible for doors to request default values to be
used for uid, gid and mode in the PnfsCreateEntryMessage request.
Default values for UID and GID are taken from the subject making the
request. Default values for mode will be derived from the parent
directory combined with a mask that depends on whether the new entry is
a directory or regular file.

Doors do not yet make use of this functionality. A follow up patch will
update the doors.

Target: trunk
Acked-by: Dmitry Litvinsev <litvinse@fnal.gov>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1956/