PnfsManager: Fix several issues with folding
PnfsManager allows collapsing or folding similar messages in the queue
together, thus avoiding querying the same information several times.

This is only possible in very specific cases. If we got two messages, A
and B, in the queue, A being before B, then processing of B can be
skipped and the reply derived from A if and only if:

- B does not have side effects

- The reply of B can be computed from the reply of A

- There are no requests in the queue in between A and B that would
  invalidate the reply of A

- The switch was ignored with Chimera

The folding technique is valuable in case of message flods caused by
aggressive retry loops in clients. In particular with PNFS we have
observed a large number of identical messages in the request queue and
folding ensures that they get processed quickly.

The current implementation had three issues:

- PnfsGetFileAttributes didn't contribute to the folding.

- Subject was ignored.

- The Message ID field was not preserved, causing issues with the DCAP
  door.


This patch replaces the folding logic to resolve those issues. The
Message base class got two methods to support folding:

invalidates() is used to check whether other messages in the queue would
change the outcome of a message. This message is as before, but the
implementation in PnfsMessage was refined.

fold() is used to let a message extract a reply from another message.
The method replaces the isIdempotent() and isSubsumedBy() methods.

The use of the fold() method ensures that Message ID is preserved.

The implementations in various PnfsMessage subclasses is updated to take
PnfsGetFileAttributes and Subject into account.

The patch cannot be applied as is to all stable branches. I will submit
a smaller patch for stable branches.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1959/