PnfsManager: Fix pathToThreadGroup in case of cache miss
pathToThreadGroup uses a cache of path prefix to PNFS DB mappings. In
case of a cache miss is returned a random thread group. This is wrong,
as it means that requests to the same path may end up on different
queues.

This patch changes the code to use an arbitrary but deterministic thread
group in case of a cache miss.

A similar problem is resolved in messageArrived.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1964/