config: set allowCleaningPreciousFiles true by default
We've agreed previously that it makes sense for pools to be allowed to remove precious 
files, even on pools with lfs set to something other than precious.

This patch enacts that decision.

RELEASE NOTES:

The property allowCleaningPreciousFiles controls whether a non-precious pool (one running 
with the lfs property set to something other than "precious") will remove a data-file if 
the corresponding namespace entry is deleted.

In previous releases of dCache, the allowCleaningPreciousFiles property had a default 
value of "false", preventing a file's data from being removed when the namespace entry is 
removed on non-precious pools.

This default behaviour has been changed.  Now, by default, deleting a file in the 
namespace will remove the corresponding data-file on the pool irrespective of the value 
of the lfs property.  The previous behaviour may be recovered by explicitly configuring 
the allowCleaningPreciousFiles property; for example, by adding the following line to 
/opt/d-cache/etc/dcache.conf

  allowCleaningPreciousFiles = false

Patch: http://rb.dcache.org/r/1966/
Acked-by: Gerd
Target: trunk