admin-interface: check cell exists before allowing cd command to succeed
The current admin cell behaviour is to allow the cd command to succeed provided the target 
cell's name is well-formed.  Although this behaviour is simple, it leads to an undesirable 
consequence: if the named cell isn't running then it's only when the user sends the next 
command that the mistake is apparent.  The user sees confusing results: the cd command 
succeeds, but the cell name was wrong.  This is a common complaint and has been commented 
on in support ticket RT #5649:

    http://rt.dcache.org/Ticket/Display.html?id=5649

This patch fixes this behaviour: it introduces a no-operation (NOOP) command that all 
cells will have.  The cd command has been updated so it checks whether the cell is running 
before succeeding.  If the user specifies a cell that doesn't exist then an error message 
is returned.

RELEASE NOTES:

The behaviour of the cd command in the admin interface has changed.  Previously, if the 
command's target cell is not running then the cd command will succeed; however, subsequent 
commands would fail will a NoRouteToCellException error message.  The new behaviour is 
that the cd command will check that the target cell is running.  If the cell is running 
then the cd command will behave as before; if the cell isn't running then the cd command 
will fail and an error message is returned.