gplazma: add real plugin loader
This patch adds a gPlazma PluginLoader class that uses the XML files store at 
/META-INF/gplazma-plugins.xml within JAR files to discover available plugins.

A previous gPlazma plugin-loader patch established PluginMetadata as the class containing 
static metadata about plugins; for example, the list of names for this plugin, the class that 
implements the plugin, etc.

This patch introduces the PluginRepository class as a container class for PluginMetadata.  The 
process of discovering the available plugins is the process of creating PluginRepository 
object.  Such an object will contain a PluginMetadata object for each plugin that has been 
discovered.

PluginRepository objects are created by factory classes; i.e., classes that implement the 
PluginRepositoryFactory interface.  The patch introduces two PluginRepositoryFactory classes: 
StaticClassPluginRepositoryFactory and XmlResourcePluginRepositoryFactory.

The StaticClassPluginRepositoryFactory is the result of refactoring the existing code from 
StaticClassPluginLoader.  This is a factory that produces PluginRepository objects based on a 
static list of GPlazmaPlugin classes (as supplied to the factory's constructor).  As with 
StaticClassPluginLoader, this class is meant to facilitate unit-testing and as demonstration 
code.

The XmlResourcePluginRepositoryFactory is the "real" PluginRepositoryFactory.  It creates 
PluginRepository objects based on scanning the calling thread's ClassLoader for resources 
located at /META-INF/gplazma-plugins.xml.  It creates a PluginRepository that contains 
PluginMetadata objects obtained by parsing these XML resources.

The common behaviour of any PluginLoader is provided by the abstract class 
AbstractPluginLoader.  Concrete subclasses of this class need only provide a 
PluginRepositoryFactory object to provide a complete implementation of PluginLoader.

Two concrete subclasses of AbstractPluginLoader are provided with this patch: 
StaticClassPluginLoader and XmlResourcePluginLoader.  They each provide the necessary 
PluginRepositoryFactory: StaticClassPluginRepositoryFactory and 
XmlResourcePluginRepositoryFactory respectively.

Both the StaticClassPluginLoader and the XmlResourcePluginLoader provide static methods for 
creating new instances: the newPluginLoader method.  Typical usage is:

  PluginLoader loader = XmlResourcePluginLoader.newPluginLoader();

  // init must be called precisely once
  loader.init();

  // new plugin with no arguments
  GPlazmaPlugin foo = loader.newPluginByName("foo");

  // new plugin with with arguments
  GPlazmaPlugin bar = loader.newPluginByName("bar", new String[]{"arg-1","arg-2"}); 

Also included are unit tests for the functionality.  They have 90%--100% code coverage; the 
notable holes in the testing is in how the code handles badly written GPlazma plugins.  
Included is a custom ClassLoader implementation to allow the testing of code that reads XML 
data from resources discovered from the calling Thread's class-loader.

Target: trunk
Patch: http://rb.dcache.org/r/1962/
Acked-by: Timur