gplazma: fix dcap door
There is a specific subclass of CacheException, PermissionDeniedCacheException, that 
should be thrown if the user is not allowed to log in.  A comment in UnionLoginStrategy 
mentions that existing code can also throw IllegalArgumentException to indicate the 
current user is not allowed to log in.

With current trunk, the plain dcap door doesn't work.  Attempts to use the door fail with 
a message like:

14 Jun 2010 17:47:09 (DCap-fandango-<unknown>-107) [] toclient(println) : 1 0 client 
failed 10015 "internalError : java.lang.IllegalArgumentException: Subject is not 
supported by GplazmaLoginStrategy"

The gPlazma cell throws an IllegalArgumentException, which is expected as an empty 
Subject is sent by the door.  The UnionLoginStrategy within the door should then wrap 
this and allow "anonymous" login.  Unfortunately, the UnionLoginStrategy propagates this 
error, resulting in the dcap command failing.

This problem is the result of two bugs: one in RemoteLoginStrategy and another in 
UnionLoginStrategy.

The bug in RemoteLoginStrategy is in how it handles the case when the remote cell throws 
an IllegalArgumentException.  If this happens, the local call will throw the (generic) 
CacheException with return-code CacheException.INVALID_ARGS.  This is clearly different 
from the two allowed cases: PermissionDeniedCacheException or IllegalArgumentException.

The fix is to catch the CacheException with CacheException.INVALID_ARGS and map it to one 
of the allowed cases.  The patch uses PermissionDeniedCacheException as throwing 
IllegalArgumentException is legacy behaviour.

The bug in UnionLoginStrategy is that it doesn't catch CacheException.  The JavaDoc for 
LoginStrategy says that implementing classes may throw PermissionDeniedCacheException 
when the login method is called to indicate the user is not allowed to log in.  However 
this method may also throw CacheException if the login fails for some other reason.  The 
UnionLoginStrategy should try all registered LoginStrategy objects and then try to log-in 
the user as an anonymous user.  By not catching CacheException, the UnionLoginStrategy 
will fail if a LoginStrategy fails for some reason other than the user isn't allowed to 
log in.

The fix is to catch CacheException too.  This is done separately from catching 
PermissionDeniedCacheException so that it can be better logged.

The patch also adds debugging to UnionLoginStrategy, which may be useful in the future.

Patch: http://rb.dcache.org/r/1993/
Acked-by: Jan
Target: trunk