poolmanager: better handling of user mistyping a command
RT ticket #5648

  http://rt.dcache.org/Ticket/Display.html?id=5648

We believe this was due to the user mistyping the "pm set" command; something 
like:

  pm set spacecostfactor=0

without the initial "-" of -spacecostfactor.  Unfortunately, the "pm set" 
command has a side-effect of creating partitions that don't already exist.  
There's also no feedback, so it isn't immediately obvious what went wrong.

The other unfortunate thing is that persistence of pool-manager configuration 
is achieved by saving a set of commands to replay when pool-manager starts up 
and this persistence relies on this side-effect.

So, this patch does three related things:

1) it changes "pm set" so the command replies with a message if it creates a 
new partition.  The UniversalSpringCell's execution of the setupFile option 
will ignore this message but interactive users will see it.  In the scenario 
described in RT #5648, that message would hopeful serve as a hint that they've 
done something wrong.

2) it introduces a new command "pm create" that allows users to create a new 
partition.  This command becomes the preferred way to create a new partition.

3) it changes the persistence so that a "pm create" command is stored before 
the "pm set" command.  Users that type "save" will then save their PoolManager 
configuration with the new "pm create" command, moving away from relying on the 
"pm set" side-effect.

There's the anticipation that, ultimately, support for the "pm set" command 
creating partitions will be removed, but this would likely need to be 
coordinated with the (few) sites using partitions.

FOR RELEASE NOTES:

This release includes support for a new command in poolmanager and the 
PoolManager.conf file: "pm create".  This is the preferred way to create new 
partitions; the existing method, using the "pm set" command, continues to work 
so the currently release is backwards compatible so existing PoolManager.conf 
files.

Future versions of dCache may remove that support, so we recommend that sites 
migrate to using "pm create".  To migrate the PoolManager.conf, either type 
"save" in the poolmanager (through the admin interface) or update 
PoolManager.conf and add a line "pm create <partition>" before the first "pm 
set <partition>" command for each partition name ("<partition>").  You don't 
need to create the default partition.

Patch is for trunk.