gplazma: don't use UTF-8 characters in source code
The PAM login strategy uses nice em-dashes in the JavaDoc, which are encoded using 
UTF-8:

http://www.fileformat.info/info/unicode/char/2014/index.htm

Unfortunately, the Java compiler doesn't default to using UTF-8 when reading the 
source-code, but (at least on Hudson) the default is (7-bit) ASCII.  UTF-8 is 
compatible with ASCII for ASCII-encodable characters, so this mostly doesn't matter.  
It's only when the source code includes non-ASCII characters that we get into problems.  
For the em-dashes, we see warning messages from the compiler about the non-mappable 
characters:

http://svn.dcache.org/build/job/dCache-Build/3263/console

This is because the UTF-8 encoding uses characters with the 8th bit set, which isn't 
allowed in ASCII.

This patch fixes these warnings by replacing the em-dashes with the less elegant (but 
ASCII encodable) hyphen symbol.

Patch: http://rb.dcache.org/r/1997/
Target: trunk
Acked-by: Timur