doors: Rely on PnfsManager to determine uid, gid and mode of new entries
This patch updates the FTP doors, xrootd door, webdav door, DCAP door
and SRM to rely on PnfsManager for determining the uid, gid and mode of
new name space entries.

The consequences are:

- Avoid some roundtrips between the door and the PnfsManager
- Unify the semantics between those doors

RELEASE NOTES: Since the semantics are unified, the semantics changes 
for some of the doors:

FTP: Previously it always used 0755 and 0644 for new entries. Now the
mode is inherited from the parent directory (with x bit cleared for
files).

SRM: Unchanged.

WebDAV: Mostly unchanged, but umask has changed.

Xrootd: Mostly unchanged, but umask has changed.

DCAP: Since DCAP allows the client to specify uid, gid and mode.
Therefore the semantics in the DCAP door is slightly different: If
provided by the client, then those values are used. If not, then
PnfsManager defaults are used. In the old semantics the door would fall
back to hardcoded values for mode if not provided by the client. After
this patch the PnfsManager would derive the mode from the mode of the
parent directory.

Target: trunk
Acked-by: Dmitry Litvinsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/1957/