PnfsManager: Fix ArrayIndexOutOfBoundsException
A recent patch to PnfsManagerV3.pathToThreadGroup introduced a bug that
could cause the returned thread group to be negative. This is caused by
the fact that hashCode() may return a negative value.

This patch adds a call to compute the absolute value of the hash code.
This is how we treat the hash code in other places on PnfsManager.

Target: trunk
Target: 1.9.9 (proposed)
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2019/