login: Fix handling of exceptions
Commit r14096 fixed a number of problems with how
IllegalArgumentException is handled in UnionLoginStrategy and
RemoteLoginStrategy. Unfortunately there was a misunderstanding of the
semantics of IllegalArgumentException.

This patch documents the semantics of IllegalArgumentException in the
LoginStrategy interface. The patch updates RemoteLoginStrategy to
rethrow INVALID_ARGS as IllegalArgumentException rather than
PermissionDeniedCacheException.


Commit r14096 also modified UnionLoginStrategy to fall through in case a
LoginStrategy throws a CacheException other than
PermissionDeniedCacheException. However since such CacheExceptions are
supposed to be transient (timeouts, communication failures, etc),
falling through could result in a user to be mapping to the wrong
Subject. E.g. a user that was supposed to have a particular UID/GID
combination ends up being treated as anonymous. Thus from the outside an
internal communication failure may appear as newly uploaded files having
the wrong owner. In the interest of fail-fast design, this patch reverts
the change and lets such errors be propagated to the door.

Target: trunk
Target: 1.9.9 (proposed)
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2021/