gPlazma 2.0: Implement main GPlazma class that implements public LoginReply login(Subject subject) throws AuthenticationException method.
The class can be instantiated with an instance of configuration or could be given a file from which the configuration is to be loaded.
Is case of of the file, the file modification time is remembered and if file is modified after it was last read, the configuration is re-read. 

The class uses XmlResourcePluginLoader to find the plugins specified in configuration
and StrategyFactory to find the implementations of the strategies.

The new utility class Strings is added to commons. this class provides a method for splitting the argument line string into arguments 
in way that supports single and double quotes around space containing arguments.

Junit Tests are added for Strings and for GPlazma

http://rb.dcache.org/r/2018/
Signed-off by: Jan, Paul