cells and srm: Establish sending CDC on cell callback
The srm does support a diagnostic context, however unfortunately we
often see session strings like this:

v1:srmcopy:116043581:-2146125640:-2146125639:-2146125639:-2146125639:-2146125639

There are only suppoesd to be three integers, but a bug in how the
diagnostic context is updated in case of nested jobs causes additional
integers to be appended. The bug is related to asynchronous callbacks
from cells. This patch resolves that issue.

When using sendMessage with a callback, I find it makes more sense to
reestablish the original CDC rather than creating a new CDC based on the
reply. The current behavior tends to make log messages generated from
the callback appear as if they are triggered by a request from the
remote host.

The SRM diagnost context (the JDC) is cleaned up and simplified. Each
Job now explicitly pushes its identifiers to the JDC rather than have
this happen indirectly in the Scheduler. The SRM specific session and
scheduler MDC entries have been removed as they did not appear to be
used and only constributed to the complexity of the JDC.

The NDC in the SRM is changed to show the current session ID. That is
what we do in the rest of dCache.

Target: trunk
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/2032/