srm: Merge the hosting cell and srm cell into one
Since the introduction of Jetty, the SRM consisted of two cells: One
cell composed with Spring for hosting Jetty and for instantiating the
SRM Storage class, and another implemented by the SRM Storage class.
This duplication lead to problems with logging as some threads were
associated with one cell and other threads with another cell.

This patch merges the two cells into one. The Storage class no longer
inherits from AbstractCell. Instead it is an AbstractCellComponent. It
is still instantiated by Spring, but is now part of the other cell.

The patch is minimal in the sense that as few components as possible
were moved to Spring. In a subsequent patch much of the internal
structure of the Storage class is moved to Spring for wiring. The
present patch however only moves the components that have to be moved to
Spring for access to cell communications.

RELEASE NOTES: The jsrm cell is no longer instantiated. The commands and 
logging specific to the jsrm cell is now available as part of the regular
SRM cell. As a consequence the output of the info command in the SRM cell 
has changed slightly. Third party monitoring scripts that parse the output 
may need to be adjusted.

Target: trunk
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/2042/