srm: Allow SRM to use synchronous replies for fast requests
The SRM protocol allows the server to use both synchronous and
asynchronous request processing. dCache always uses asynchronous replies
(except for ls, which can be configured to be either synchronous or
asynchronous). We can assume that clients support both types of replies,
as other SRM implementations (e.g. Bestman) use synchronous replies in
some situations.

The asynchronous reply has the benefit of immediately freeing up the
request thread and the socket. The downside is that it forces the client
to poll for the result. This adds latency to the request because the
client sleeps for some seconds before polling for the result, and it
adds additional processing overhead, as the client needs to establish a
new GSI connection (which needs to be authenticated and will cause
credentials to be delegated).

This patch allows the SRM to switch between synchronous and asynchronous
replies depending on how quickly the request can be processed: If
request processing finishes within a configurable amount of time then a
synchronous reply is provided. Otherwise an asynchronous reply is
generated. If the limit is set to 0, then asynchronous replies will
always be used. If set to infinity (implemented as Long.MAX_VALUE, which
for all practical purposes equals infinity) then synchronous replies
will always be used.

The patch only influences SRM v2 requests. SRM v1 request processing is
unchanged.


A more advanced version of the logic could change the delay dynamically
based on the current number of TCP connections to the SRM service. As
the current patch is quite useful in itself, I suggest we implement a
more advanced version in a separate patch.


Notice that I as far as possible have tried to copy the naming scheme of
existing variables for the various request types
({Get,Put,BringOnline}Req, LsRequest, or no Req/Request suffix),
capitalization and hyphenization. I avoided abreviations.


RELEASE NOTES: The output format of the info command of the SRM changed
in the following way: A typo in the section for put scheduler parameters
was fixed. Information about the setting for asynchronous ls was
removed. Information about the asynchronous crossover time was added for
get, put, ls and bring online.

RELEASE NOTES: The "set async ls" command was removed. The commands "set
switch to asynchronous mode delay get", "set switch to asynchronous mode
delay put", "set switch to asynchronous mode delay bringonline" and "set
switch to asynchronous mode delay ls" were added.

RELEASE NOTES: The configuration parameter srmAsynchronousLs was
deprecated. The configuration parameters
srmLsRequestSwitchToAsynchronousModeDelay,
srmGetReqSwitchToAsynchronousModeDelay,
srmPutReqSwitchToAsynchronousModeDelay, and
srmBringOnlineReqSwitchToAsynchronousModeDelay were added.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/2020/