commons: update unit-tests for FQAN class
In general, FQANs have a specific structure.  They have a hierarchical group structure, 
an optional "Role" attribute and an optional "Capability" attribute.  Stopping after 
descending once in the group hierarchy, an FQAN may have one of the following forms:

  /a
  /a/Role=r
  /a/Capability=c
  /a/Role=r/Capability=c
  /a/b
  /a/b/Role=r
  /a/b/Capability=c
  /a/b/Role=r/Capability=c

where a, b, r and c are (more-or-less) arbitrary strings.

The current unit-tests for the FQAN class provide a very minimal check.  They test only 
that an FQAN with no subgroup or capability but with a role.  They check such that 
equals is reflexive and that, when the role value is changed, the equals fails.

The current tests are also limited since there is no tests for the methods that allow 
one to obtain information from the class.

This patch adds tests that verify:

  1. the equals for all above forms of FQAN are reflexive.

  2. that no two of the above forms are equal with any of the other forms.

  3. that a change in any element's data (a, b, r or c in above example) will result 
     FQANs with the changed element being different from all original FQANs.

  4. call to getGroup returns /a or /a/b (as appropriate)

  5. call to getRole for FQANs with /Role=r returns r

  6. call to getRole for FQANs without /Role=r returns an empty string.

  7. call to getCapability for FQANs with /Capability=c returns c

  8. call to getCapability for FQANs without /Capability=c returns an empty string.

Patch: http://rb.dcache.org/r/2044/
Acked-by: Timur
Target: trunk