commons: add hasRole and hasCapability methods to FQAN
There are some places where code needs to know if an FQAN has a role.  Rather than 
testing against specific hard-coded values (like an empty string or null) as is done 
currently, the code should be able to ask the FQAN object for this information directly.

This patch adds the method hasRole.  This method returns true if the FQAN has a 
non-empty Role and returns false otherwise.  A future patch will update other code to 
use this new method.

The patch also adds the method hasCapability for completeness.

Four tests are added to the unit-tests.  These test the positive and negative replies of 
each of the methods.

Target: trunk
Patch: http://rb.dcache.org/r/2053/
Acked-by: Gerd