Ant-build: Add new server.deb buildfile
The patch changes the ant build for the debian server package to using
the standard debhelper procedure instead of just rpm conversion with
alien. In the long run this can serve to produce higher quality debian
packages and to make add on packages that configure parts of the system.

Acked-by: Gerd Behrmann <behrmann@ndgf.org>
Patch: http://rb.dcache.org/r/2040/
Target: trunk