srm: remover deadlocks related to Request.getUser calls coming from scheduler.stateChanged(), remove other possible deadlocks by converting some of the variables to final. Tigran recently observed SRM deadlocks. All of the deadlocks were caused by the simultaneous changes of file request by the scheduler / srm execution of the request and simulteneous client operations on the same request (like AbortRequest). When the status of the file request is changed the scheduler is notified using scheduler.stateChanged() for accounting purposes. Inside the implementation of the stateChanged, the scheduler gets the user of the request, access to user was protected by the locks owned by the request which is a container of the file request. This was locking on request level lock inside the code locked on the file request lock. the Abort and other state changinging requests are locking on the request followed by the locking on the file request.
In order to avoid deadlock, the user of the request is made a final field and its get method does not use locking. In general the rule for locking on srm requests should be always ordered, first lock on request, then on file request. Do not call any of the request level locking code from the file request code that is locked.
(Ordering of the locks).

This patch also converts some other request level variables to final and removes locking on their access. It also corrects some of the formatting in the affected code.

http://rb.dcache.org/r/2046/
Signed-off by: Gerd, Tigran