FQAN: add support for /Role=NULL
For better or worse, dCache accepts a FQAN-role with a special value: "NULL".  This was 
used to indicate that the FQAN has no role, so that the two FQANs "/atlas/Role=NULL" and 
"/atlas" are equivalent.

Current support for this is somewhat sporadic, focused in the gPlazma classes.

This patch introduces this special behaviour into FQAN: if an FQAN is specified with 
/Role=NULL or /Capability=NULL then the FQAN is automatically converted to its nominal 
form.

Unit tests are added to confirm correct behaviour.

Patch: http://rb.dcache.org/r/2064/
Acked-by: Thomas
Target: trunk