FQAN: add isValid method to test if a string is a valid FQAN
An FQAN is not an arbitrary string: it has a specific format, as described in:

    http://edg-wp2.web.cern.ch/edg-wp2/security/voms/edg-voms-credential.pdf

This patch adds support for testing whether some arbitrary string is a valid FQAN.  The 
isValid static method returns true if the string may be represented as an FQAN, false 
otherwise.

The patch also adds unit tests to verify correct behaviour: both in accepting valid FQANs 
and rejecting invalid FQANs.

Acked-by: Gerd
Patch: http://rb.dcache.org/r/2067/
Target: trunk